requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
transformers==4.46.3
torch==2.5.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.1.2
openpyxl==3.1.5
Pillow==11.0.0